fixed redirection when user tries to access parrts they dont have rights to
